Renamed bundle_update_service to just exporter.
